requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
flask>=2.3.0
flask-cors>=4.0.0
pillow>=10.0.0
scikit-learn>=1.3.0
seaborn>=0.12.0
opencv-python>=4.8.0
onnx>=1.15.0
onnxruntime>=1.16.0 